requirements2.txt
colour==0.1.5
coverage==4.5.3
Django==1.11
django-bootstrap3==11.0.0
django-cleanup==3.2.0
django-cors-headers==3.0.2
django-easy-select2==1.5.6
djangorestframework==3.9.4
Pillow==6.0.0
python-gettext==4.0
pytz==2019.2
django-modeltranslation==0.13.3
django-colorfield==0.1.15
dropbox==9.4.0
django-storages==1.8
django-dbbackup==3.2.0
celery==3.1.25
redis==2.10.3
django-celery==3.3.1
gunicorn==20.0.4
django-monthfield==0.1.3
pymysql==0.9.3
xlrd==1.2.0


#python==3.7.4
#virtualenv==16.7.3